requirements.txt
feedparser==6.0.11
nba_api>=1.4.1
openai>=1.30.0
pandas>=2.0.0
PyYAML==6.0.2
python-dotenv==1.0.1
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
